Marker component added
